Refactor SessionsController#destroy
